# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jeonbuk Career Due-Diligence Backend",
    "description": "Audits a metropolitan-area posting against curated Jeonbuk postings using evidence-grounded field extraction, converts unresolved gaps into verification actions, and runs a deterministic hypothetical cash comparison. Not a company-rating or future-prediction service.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/postings/analyze": {
      "post": {
        "tags": [
          "postings"
        ],
        "summary": "Analyze",
        "operationId": "analyze_api_v1_postings_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostingInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostingAnalysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/postings/match": {
      "post": {
        "tags": [
          "postings"
        ],
        "summary": "Match",
        "operationId": "match_api_v1_postings_match_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/postings/home-region-matches": {
      "post": {
        "tags": [
          "postings"
        ],
        "summary": "Home Region Matches",
        "description": "Real curated-pair candidates for one capital-area posting, scoped to\nthe server-configured home region only -- the request has no field a\nclient could use to ask for a different region.",
        "operationId": "home_region_matches_api_v1_postings_home_region_matches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HomeRegionMatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/postings/analyze-by-id": {
      "post": {
        "tags": [
          "postings"
        ],
        "summary": "Analyze By Id",
        "description": "Resolves posting_id -> private full text server-side (never returned\nto the client raw), then picks one of two pipelines:\n\n  - A `Work24StructuredPosting` fixture exists for this posting_id ->\n    the hybrid pipeline (`analyze_posting_hybrid`): salary/employment_type\n    come deterministically from that record, the remaining four fields\n    from an sLLM's free-text extraction.\n  - No such record exists (every posting outside the current tiny\n    research-persona demo set) -> the exact same\n    `analyze_posting()` six-field pipeline as /postings/analyze,\n    completely unchanged from before this feature existed.\n\nRaises PrivateDataUnavailableError (503) if this machine has no private\ncopy for posting_id right now, in either case.",
        "operationId": "analyze_by_id_api_v1_postings_analyze_by_id_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeByIdRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostingAnalysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalyzeByIdRequest": {
        "properties": {
          "posting_id": {
            "type": "string",
            "minLength": 1,
            "title": "Posting Id"
          },
          "expected_occupation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Occupation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "posting_id"
        ],
        "title": "AnalyzeByIdRequest"
      },
      "AuditedField": {
        "properties": {
          "field": {
            "type": "string",
            "enum": [
              "salary",
              "duties",
              "tools_or_skills",
              "training_or_mentoring",
              "probation_terms",
              "employment_type"
            ],
            "title": "Field"
          },
          "status": {
            "type": "string",
            "enum": [
              "confirmed",
              "vague",
              "absent"
            ],
            "title": "Status"
          },
          "evidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvidenceSpan"
              },
              {
                "type": "null"
              }
            ]
          },
          "reason_code": {
            "type": "string",
            "minLength": 1,
            "title": "Reason Code",
            "description": "Deterministic rule code explaining the status decision."
          },
          "provenance": {
            "type": "string",
            "enum": [
              "WORK24_STRUCTURED",
              "SLM_EXTRACTED",
              "USER_REPORTED"
            ],
            "title": "Provenance",
            "description": "Where this field's value actually came from: WORK24_STRUCTURED (고용24 등록 정보, deterministic, never LLM-touched), SLM_EXTRACTED (공고 본문 자유서술을 분석), or USER_REPORTED (not produced anywhere yet)."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "field",
          "status",
          "reason_code",
          "provenance"
        ],
        "title": "AuditedField"
      },
...
      "EvidenceSpan": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "start": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Start"
          },
          "end": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "End"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "start",
          "end"
        ],
        "title": "EvidenceSpan",
        "description": "An exact, offset-verifiable quote from the source posting text."
      },
      "ExternalContext": {
        "properties": {
          "source": {
            "type": "string",
            "minLength": 1,
            "title": "Source"
          },
          "reference_date": {
            "type": "string",
            "title": "Reference Date",
            "description": "ISO-8601 date the external figure was published or fetched."
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "title": "Description"
          },
          "limitations": {
            "type": "string",
            "minLength": 1,
            "title": "Limitations"
          },
          "values": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string"
                }
              ]
            },
            "type": "object",
            "title": "Values"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "reference_date",
          "description",
          "limitations"
        ],
        "title": "ExternalContext"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HomeRegionMatchRequest": {
        "properties": {
          "metro_posting_id": {
            "type": "string",
            "minLength": 1,
            "title": "Metro Posting Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "metro_posting_id"
        ],
        "title": "HomeRegionMatchRequest",
        "description": "Client supplies only which capital-area posting it's looking at --\nnever a region. The home region is always the server-configured value\n(see app.datasets.loader.home_region()); there is no field here a\nclient could set to request a different one."
      },
      "MatchCandidate": {
        "properties": {
          "posting_id": {
            "type": "string",
            "title": "Posting Id"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "region": {
            "type": "string",
            "title": "Region"
          },
          "municipality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Municipality"
          },
          "occupation": {
            "type": "string",
            "title": "Occupation"
          },
          "employment_type": {
            "type": "string",
            "title": "Employment Type"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "source_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Text",
            "description": "Posting text when the configured source permits runtime display; may be null."
          },
          "is_synthetic": {
            "type": "boolean",
            "title": "Is Synthetic",
            "default": false
          },
          "matching_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matching Fields"
          },
          "mismatch_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Mismatch Fields"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Plain statement that this candidate comes from a finite curated dataset, not a claim of best fit."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "posting_id",
          "region",
          "occupation",
          "employment_type",
          "matching_fields",
          "mismatch_fields",
          "description"
        ],
        "title": "MatchCandidate"
      },
      "MatchRequest": {
        "properties": {
          "occupation": {
            "type": "string",
            "minLength": 1,
            "title": "Occupation"
          },
          "employment_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employment Type"
          },
          "posting_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posting Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "occupation"
        ],
        "title": "MatchRequest"
      },
      "MatchResponse": {
        "properties": {
          "query_occupation": {
            "type": "string",
            "title": "Query Occupation"
          },
          "query_employment_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Employment Type"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/MatchCandidate"
            },
            "type": "array",
            "title": "Candidates"
          },
          "dataset_description": {
            "type": "string",
            "title": "Dataset Description"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query_occupation",
          "candidates",
          "dataset_description"
        ],
        "title": "MatchResponse"
      },
      "PostingAnalysis": {
        "properties": {
          "posting_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posting Id"
          },
          "occupation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occupation"
          },
          "employment_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employment Type"
          },
          "fields": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AuditedField"
            },
            "propertyNames": {
              "enum": [
                "salary",
                "duties",
                "tools_or_skills",
                "training_or_mentoring",
                "probation_terms",
                "employment_type"
              ]
            },
            "type": "object",
            "title": "Fields"
          },
          "verification_actions": {
            "items": {
              "$ref": "#/components/schemas/VerificationAction"
            },
            "type": "array",
            "title": "Verification Actions"
          },
          "validation_warnings": {
            "items": {
              "$ref": "#/components/schemas/ValidationWarning"
            },
            "type": "array",
            "title": "Validation Warnings"
          },
          "external_context": {
            "items": {
              "$ref": "#/components/schemas/ExternalContext"
            },
            "type": "array",
            "title": "External Context"
          },
          "work_hours": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkHoursInfo"
              },
              {
                "type": "null"
              }
            ],
            "description": "근로시간·교대제, sourced only from a Work24StructuredPosting record -- never from an LLM. None means no such record exists for this posting (unchanged 'not_evaluated' MVP-scope behavior); see WorkHoursInfo.status for the distinction between 'confirmed' and 'structured_absent'."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "fields"
        ],
        "title": "PostingAnalysis"
      },
      "PostingInput": {
        "properties": {
          "posting_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posting Id"
          },
          "source_text": {
            "type": "string",
            "minLength": 1,
            "title": "Source Text"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "expected_occupation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Occupation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_text"
        ],
        "title": "PostingInput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationWarning": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "field": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "salary",
                  "duties",
                  "tools_or_skills",
                  "training_or_mentoring",
                  "probation_terms",
                  "employment_type"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ValidationWarning"
      },
      "VerificationAction": {
        "properties": {
          "field": {
            "type": "string",
            "enum": [
              "salary",
              "duties",
              "tools_or_skills",
              "training_or_mentoring",
              "probation_terms",
              "employment_type"
            ],
            "title": "Field"
          },
          "channel": {
            "type": "string",
            "enum": [
              "email",
              "phone",
              "interview",
              "document_review",
              "pre_contract"
            ],
            "title": "Channel"
          },
          "prompt": {
            "type": "string",
            "minLength": 1,
            "title": "Prompt"
          },
          "triggered_by_status": {
            "type": "string",
            "enum": [
              "vague",
              "absent"
            ],
            "title": "Triggered By Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "field",
          "channel",
          "prompt",
          "triggered_by_status"
        ],
        "title": "VerificationAction"
      },
      "WorkHoursInfo": {
        "properties": {
          "weekly_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekly Hours"
          },
          "detailed_work_hours": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detailed Work Hours"
          },
          "shift_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shift Type"
          },
          "status": {
            "type": "string",
            "enum": [
              "confirmed",
              "structured_absent"
            ],
            "title": "Status"
          },
          "provenance": {
            "type": "string",
            "const": "WORK24_STRUCTURED",
            "title": "Provenance",
            "default": "WORK24_STRUCTURED"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "WorkHoursInfo",
        "description": "Presentation-ready 근로시간·교대제 info, sourced *only* from\n`Work24StructuredPosting` -- never from an LLM/sLLM, never estimated.\n\n`status`:\n  - \"confirmed\": at least one of weekly_hours/detailed_work_hours/\n    shift_type is present on the structured record.\n  - \"structured_absent\": a Work24StructuredPosting record exists for\n    this posting_id, but none of those three sub-fields are stated on\n    it. Frontend must render this as \"고용24 등록 정보에서 확인되지\n    않음\" -- never \"확인 불가\" (that phrase is reserved for the\n    evidence-harness vague/absent flow, which this is not).\n\n`PostingAnalysis.work_hours is None` (the field is simply absent from\nthe response) is the *third*, different case: no Work24StructuredPosting\nrecord exists for this posting_id at all. The frontend keeps showing\n`not_evaluated` for that case, unchanged from before this feature."
      }
    }
  }
}